pyproject.toml
[tool.poetry]
name = "mec-inm"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "mec_inm"}]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "2.1"
pandasql = "^0.7.3"
openpyxl = "^3.1.2"
biopython = "^1.81"
cassandra-driver = "^3.28.0"
loguru = "^0.7.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
main = "mec_inm.main:run"
